Moved classes for day 2
